[Enhancement] support virtual column row_id and segment_id
follow #66214. Support more virtual columns

+++ be/src/storage/virtual_column_utils.h
@@ -31,7 +31,7 @@ public:
     struct Options {
         int32_t tablet_id;
         int32_t segment_id;
-        int32_t num_rows;
+        int64_t num_rows;
     };
     static StatusOr<ColumnIterator*> create_virtual_column_iterator(const Options& options,
                                                                     const std::string_view col_name);
